requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
setuptools
daphne

